refactor: improve Svelte state and routing patterns
