+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "zwfm-knabbel",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.1.0",
   "type": "module",
   "scripts": {
     "dev": "vite dev",
