+++ src/lib/components/ui/EmptyState.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import type { Component } from 'svelte'
+  import { resolveInternalHref } from '$lib/utils/routes'
 
   interface Props {
     icon: Component
@@ -9,6 +10,8 @@
   }
 
   let { icon: Icon, title, description, action }: Props = $props()
+
+  const resolvedActionHref = $derived(action ? resolveInternalHref(action.href) : undefined)
 </script>
 
 <div class="card bg-base-100">
@@ -23,7 +26,7 @@
     {/if}
     {#if action}
       <a
-        href={action.href}
+        href={resolvedActionHref}
         class="btn mt-4 btn-primary">{action.label}</a
       >
     {/if}
