+++ src/lib/components/ui/TextInput.svelte
@@ -34,8 +34,7 @@
     {id}
     {type}
     bind:value
-    class="input w-full"
-    class:input-error={error}
+    class={['input w-full', error && 'input-error']}
     {placeholder}
     {disabled}
   />
