+++ src/lib/components/ui/TextareaInput.svelte
@@ -33,8 +33,7 @@
   <textarea
     {id}
     bind:value
-    class="textarea w-full resize-y leading-relaxed"
-    class:textarea-error={error}
+    class={['textarea w-full resize-y leading-relaxed', error && 'textarea-error']}
     {rows}
     {placeholder}
     {disabled}
