+++ src/lib/types/index.ts
@@ -19,6 +19,7 @@ export interface StationConfig {
   enabled: boolean
   stationVoiceId: number | null
   mixPoint: number
+  savedMixPoint: number
   audioUrl: string | null
   hasAudio: boolean
   jingleFile: File | null
