+++ src/lib/utils/pagination.ts
@@ -15,6 +15,11 @@ export interface PaginationInfo {
   hasPrevPage: boolean
 }
 
+export interface VisiblePage {
+  id: string
+  page: number | null
+}
+
 /**
  * Parse pagination params from URL search params
  */
@@ -51,16 +56,18 @@ export function getVisiblePages(
   currentPage: number,
   totalPages: number,
   maxVisible = 7
-): (number | 'ellipsis')[] {
+): VisiblePage[] {
+  const pageToken = (page: number): VisiblePage => ({ id: `page-${page}`, page })
+
   if (totalPages <= maxVisible) {
-    return Array.from({ length: totalPages }, (_, i) => i + 1)
+    return Array.from({ length: totalPages }, (_, i) => pageToken(i + 1))
   }
 
-  const pages: (number | 'ellipsis')[] = []
+  const pages: VisiblePage[] = []
   const sideCount = Math.floor((maxVisible - 3) / 2)
 
   // Always include first page
-  pages.push(1)
+  pages.push(pageToken(1))
 
   // Calculate range around current page
   let rangeStart = Math.max(2, currentPage - sideCount)
@@ -80,21 +87,21 @@ export function getVisiblePages(
 
   // Add ellipsis before range if needed
   if (rangeStart > 2) {
-    pages.push('ellipsis')
+    pages.push({ id: 'ellipsis-left', page: null })
   }
 
   // Add range pages
   for (let i = rangeStart; i <= rangeEnd; i++) {
-    pages.push(i)
+    pages.push(pageToken(i))
   }
 
   // Add ellipsis after range if needed
   if (rangeEnd < totalPages - 1) {
-    pages.push('ellipsis')
+    pages.push({ id: 'ellipsis-right', page: null })
   }
 
   // Always include last page
-  pages.push(totalPages)
+  pages.push(pageToken(totalPages))
 
   return pages
 }
